+++ service/internal/security/errors.go
@@ -1,0 +1,19 @@
+package security
+
+const (
+	ErrCertNotFound        = Error("not found")
+	ErrCertificateEncode   = Error("certificate encode error")
+	ErrPublicKeyMarshal    = Error("public key marshal error")
+	ErrHSMUnexpected       = Error("hsm unexpected")
+	ErrHSMDecrypt          = Error("hsm decrypt error")
+	ErrHSMNotFound         = Error("hsm unavailable")
+	ErrKeyConfig           = Error("key configuration error")
+	ErrUnknownHashFunction = Error("unknown hash function")
+)
+const keyLength = 32
+
+type Error string
+
+func (e Error) Error() string {
+	return string(e)
+}
